Start speedo addition
Add initial speedometer implementation.

Documentation still needs to be completed.

A note on current speedo scripts:

Current will work on both controller only and with entry but, on with entry the script currently will not yet remove interface items on tag, and that is something I will do shortly, but since then if those edits are made then it will conflict with controller only usages, so that is why the "extras" folder has been added as a placeholder for when I will move the original speedo files to there for vehicle controller only usages at the time that I opted to make the entry only specific edits to the speedo scripts.