Write initial speedometer doku(s)
These doku(s) are the initial docs as far as vehiclecontroller only usage and these docs will eventually as is be move with pertaining scripts to extras folder as then the scripts will be modified to factor for disabling on state as such when using entry script and as such tnot only will the scripts slightly change for new changes but also the docs will change slightly thus the need to have the old moved to extras dir simply so that folks still opting to only use vehicle controller only usage and not opt for player entry addition will still be able to use the speedo even after entry related changes. Hopefully that makes sense. ;)